packet rate fixes

+++ Assets/Networking Scripts/MeshNetworkTransform.cs
@@ -190,7 +190,7 @@ public class MeshNetworkTransform : MonoBehaviour, IReceivesPacket<MeshPacket>, 
                 
 
             }
-            if(Time.time - lastBroadcastTime > (1 / BROADCAST_RATE)) {
+            if(Time.time - lastBroadcastTime > (float)(1 / BROADCAST_RATE)) {
                 lastBroadcastTime = Time.time;
                 BroadcastUpdate();
